Added the creation and use of the VkPipeline to draw a triangle
The VkPipeline holds the entire state of the GPU required to draw primitives, from vertex shading through rasterization and fragment shading.  It takes loaded shader modules and exposes them to the driver, along with the rest of the device state for drawing a batch, triggering compilation of the state.  Thus, it could fail or be an expensive operation.