More proper cleanup solution, no errors
Implemented a FILO system, recording objects on creation to be destroyed in reverse order, via a deque of functions containing destroy operations on those objects.  It seems like the deque could just as easily be a vector, but perhaps over time that changes?

Added initializers for Fence and Semaphore create info.