Animating rotating triangle using matrix constant pushed to VS
Now pushing constants to the vertex shader including a matrix by which to transform the vertex position.  This matrix is pre-built to contain the MVP.

Slightly reorganized InitPipelines() to be a bit tidier, and added comments, as it was getting pretty long.

Now it animates a rotating triangle without backface culling.  Makes sense why we setup the device without backface culling now.