Added new greyscale fragment shader, fixed binding sort
There's a new greyscale_triangle.frag that just applies a greyscale dot approximation to the incoming color value and outputs it.  This is now applied to every other mesh in the volumetric grid.

I checked the results of the sort to minimize binding and realized I was only sorting by pipeline, not pipeline and mesh.  I had 8 pipeline binds and over 1000 mesh binds.  Now we sort by both and the number of binding changes is 2 and 7, respectively,

The 20x20x20 grid still runs at around 8fps, suggesting something in the GPU.  Need metrics.