Added descriptors so camera matrix constants can be set for shaders
New engine function: CreateBuffer, using the vma allocator to create a generic buffer of a requested size and usage.

Now transforming triangle mesh vertex positions by viewproj in the trimesh vertex shader.

This required building a DescriptorSet and its related Layout and a Pool to hold it, in the new InitDescriptors().

First it creates a pool, then a layout (via binding and create info), then for each FRAME_OVERLAP it creates a GPUCameraData-sized buffer of UNIFORM type and of CPU_TO_GPU usage.

Next, it creates a descriptorset containing pointers to the common descriptor pool, above, and engine-global set layout.  Finally, it updates that descriptorset with write configuration information about our new camera-info-sized buffer.

The global set layout is attached to the meshPipelineLayoutInfo so that it knows to expect and use the descriptor data in the shaders.

The GPUCameraData struct holds the recalculated camera data each frame, which is copied into memory mapped onto the allocated camera buffer.