The VS tri_mesh.vert now gets the model matrix from a shader storage buffer and instance value instead of a push constant.
The shader storage buffer object (SSBO) contains just an array of structs containing one mat4.  They're allocated using their own descriptor set and buffer in the per-frame object.  Each frame during draw, the matrices are copied into this array.  The draw passes the "first instance" value which controls the "gl_BaseInstance" value available in the shader, which can then be used to lookup a matrix in the shader.